# infophysics_impl_grade/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: InformationPhysics API
  version: 0.1.0
  description: >
    Implementation-grade API surface for the InformationPhysics.ai platform.
    Core primitives: Information Objects (IO), versioned representations, derivation DAG,
    and the measurement workbench (SQL/retrieval/LLM/estimator instruments).
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XTenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema: { type: string }
      description: Tenant partition key (enforced by auth + RLS).
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    ISODateTime:
      type: string
      format: date-time
    UUID:
      type: string
      format: uuid

    # ---- Core types ----
    InformationObject:
      type: object
      required: [io_id, tenant_id, type, created_at]
      properties:
        io_id: { $ref: "#/components/schemas/UUID" }
        tenant_id: { type: string }
        type: { type: string, example: Document }
        created_at: { $ref: "#/components/schemas/ISODateTime" }
        captured_at: { $ref: "#/components/schemas/ISODateTime" }
        raw:
          type: object
          properties:
            raw_uri: { type: string, example: "s3://infophysics/tenantA/raw/abc.pdf" }
            raw_hash: { type: string, example: "sha256:..." }
            mime_type: { type: string, example: "application/pdf" }
            size_bytes: { type: integer, example: 123456 }
        context:
          type: object
          properties:
            source_system: { type: string, example: "ACC" }
            source_object_id: { type: string, example: "acc:projects/123/docs/456" }
            author: { type: string, example: "Jane Smith" }
            time_range:
              type: object
              properties:
                start: { $ref: "#/components/schemas/ISODateTime" }
                end: { $ref: "#/components/schemas/ISODateTime" }
            permissions_ref: { type: string, example: "policy:tenantA:projects:123" }
            policy_scope_id: { type: string, example: "projects_123" }
        representations:
          type: object
          properties:
            extracted_text_versions: { type: array, items: { type: string } }
            chunk_versions: { type: array, items: { type: string } }
            embedding_versions: { type: array, items: { type: string } }
            summary_versions: { type: array, items: { type: string } }
            view_versions: { type: array, items: { type: string } }
        lineage:
          type: object
          properties:
            lineage_root: { $ref: "#/components/schemas/UUID" }
            derivations: { type: array, items: { $ref: "#/components/schemas/UUID" } }
...
    CreateIORequest:
      type: object
      required: [type, raw, context]
      properties:
        type: { type: string, example: Document }
        raw:
          type: object
          description: >
            Either provide an upload workflow (pre-signed URL) or an existing raw_uri+hash.
          properties:
            upload_url: { type: string, nullable: true }
            raw_uri: { type: string, nullable: true }
            raw_hash: { type: string, nullable: true }
            mime_type: { type: string, nullable: true }
            size_bytes: { type: integer, nullable: true }
        context:
          type: object
          properties:
            source_system: { type: string, nullable: true }
            source_object_id: { type: string, nullable: true }
            author: { type: string, nullable: true }
            time_range:
              type: object
              properties:
                start: { $ref: "#/components/schemas/ISODateTime" }
                end: { $ref: "#/components/schemas/ISODateTime" }
            permissions_ref: { type: string, nullable: true }
            policy_scope_id: { type: string, nullable: true }
        dedupe:
          type: object
          properties:
            mode:
              type: string
              enum: [none, hash, source_object_id, hash_or_source]
              default: hash_or_source
    CreateIOResponse:
      type: object
      required: [item]
      properties:
        item: { $ref: "#/components/schemas/InformationObject" }
...
security:
  - bearerAuth: []
paths:
...
  /v1/io:
    post:
      tags: [IO]
      summary: Capture an Information Object (preserve-first)
      parameters: [ { $ref: "#/components/parameters/XTenantId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateIORequest" }
            examples:
              contract_pdf:
                summary: Capture a contract PDF from object storage
                value:
                  type: Document
                  raw:
                    raw_uri: "s3://infophysics/tenantA/raw/contracts/msa.pdf"
                    raw_hash: "sha256:7df4..."
                    mime_type: "application/pdf"
                    size_bytes: 842113
                  context:
                    source_system: "SharePoint"
                    source_object_id: "sp:sites/legal/contracts/msa.pdf"
                    author: "Legal Ops"
                    permissions_ref: "policy:tenantA:legal"
                    policy_scope_id: "legal"
                  dedupe:
                    mode: "hash_or_source"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateIOResponse" }
        "400": { description: Bad Request, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: Unauthorized }

    get:
      tags: [IO]
      summary: List Information Objects
      parameters:
        - { $ref: "#/components/parameters/XTenantId" }
        - name: type
          in: query
          schema: { type: string }
        - name: source_system
          in: query
          schema: { type: string }
        - name: created_after
          in: query
          schema: { $ref: "#/components/schemas/ISODateTime" }
        - name: created_before
          in: query
          schema: { $ref: "#/components/schemas/ISODateTime" }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/InformationObject" } }
...